+++ src/Math/Math.cpp
@@ -77,19 +77,19 @@ void Tree::setWeight() {
     }
 }
 
-double Math::variance(VectorXd v) {
+double Math::variance(VectorXf v) {
     double mean = v.mean();
     v = (v.array() - mean).matrix();
     return v.squaredNorm() / v.size();
 }
 
-double Math::covariance(VectorXd v1, VectorXd v2) {
+double Math::covariance(VectorXf v1, VectorXf v2) {
     double m1 = v1.mean();
     double m2 = v2.mean();
     return ((v1.array() - m1) * (v2.array() - m2)).matrix().mean();
 }
 
-double Math::correlation(VectorXd v1, VectorXd v2) {
+double Math::correlation(VectorXf v1, VectorXf v2) {
     double cov = covariance(v1, v2);
     double var1 = std(v1);
     double var2 = std(v2);
@@ -99,7 +99,7 @@ double Math::correlation(VectorXd v1, VectorXd v2) {
     return cov / (var1 * var2);
 }
 
-double Math::std(VectorXd v) {
+double Math::std(VectorXf v) {
     double mean = v.mean();
     v = (v.array() - mean).matrix();
     return sqrt(v.squaredNorm() / v.size());
@@ -115,7 +115,7 @@ void Math::removeCol(MatrixXf *mptr, long y) {
     mptr->conservativeResize(numRows, numCols);
 }
 
-VectorXd Math::L2Thresholding(VectorXd in) {
+VectorXf Math::L2Thresholding(VectorXf in) {
     if (in.norm()>1){
         return in/in.norm();
     }
@@ -188,7 +188,7 @@ void Math::updateMap(unordered_map<long, treeNode *> *mptr, minXY xy) {
 MatrixXf Math::appendColRow(MatrixXf mat, minXY xy) {
     long r = mat.rows();
     MatrixXf result = MatrixXf::Zero(r + 1, r + 1);
-    VectorXd col = VectorXd::Zero(r);
+    VectorXf col = VectorXf::Zero(r);
     col = ((mat.col(xy.x).array()).max(mat.col(xy.y).array())).matrix();
     result.block(0, 0, r, r) = mat;
     result.block(0, r, r, 1) = col;
